refactor: dendrogram drawers are now in the Matplotlib and Cairo submodules
